+++ src/components/Layout/Layout.tsx
@@ -1,0 +1,12 @@
+import { Outlet } from 'react-router-dom';
+import Header from '../Header/Header';
+
+
+export default function Layout() {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+}
