+++ src/components/PrivateRoute/PrivateRoute.tsx
@@ -7,7 +7,7 @@ type PrivateRouteProps = {
 export default function PrivateRoute({
   children,
 }: PrivateRouteProps): JSX.Element {
-  const hasAccess = false;
+  const hasAccess = true;
 
   return hasAccess ? children : <Navigate to={'/login'} />;
 }
