+++ src/mocks/offers.ts
@@ -1,0 +1,45 @@
+export const Setting = {
+  offersCount: 312,
+  places: [
+    {
+      id: '1',
+      title: 'Beautiful & luxurious apartment at great location',
+      price: 120,
+      type: 'Apartment',
+      isPremium: true,
+      isFavorite: false,
+    },
+    {
+      id: '2',
+      title: 'Wood and stone place',
+      price: 80,
+      type: 'Room',
+      isPremium: false,
+      isFavorite: true,
+    },
+    {
+      id: '3',
+      title: 'Canal View Prinsengracht',
+      price: 132,
+      type: 'Apartment',
+      isPremium: false,
+      isFavorite: false,
+    },
+    {
+      id: '4',
+      title: 'Nice, cozy, warm big bed apartment',
+      price: 180,
+      type: 'Apartment',
+      isPremium: true,
+      isFavorite: false,
+    },
+    {
+      id: '5',
+      title: 'Wood and stone place',
+      price: 80,
+      type: 'Room',
+      isPremium: false,
+      isFavorite: true,
+    },
+  ],
+};
