+++ src/pages/login/login.tsx
@@ -1,3 +1,5 @@
+import Logo from '../../components/Logo/Logo';
+
 export default function Login() {
   return (
     <div className="page page--gray page--login">
@@ -5,15 +7,7 @@ export default function Login() {
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="main.html">
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width={81}
-                  height={41}
-                />
-              </a>
+              <Logo />
             </div>
           </div>
         </div>
