+++ src/pages/main/main.tsx
@@ -1,4 +1,4 @@
-import PlaceCard from '../../components/place-card/place-card';
+import OffersList from '../../components/OffersList/OffersList';
 import { Place } from '../../const';
 
 type MainProps = {
@@ -9,44 +9,6 @@ type MainProps = {
 export default function Main({ places, offersCount }: MainProps) {
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link header__logo-link--active">
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width={81}
-                  height={41}
-                />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                      Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
@@ -118,18 +80,7 @@ export default function Main({ places, offersCount }: MainProps) {
                   </li>
                 </ul>
               </form>
-              <div className="cities__places-list places__list tabs__content">
-                {places.map((place) => (
-                  <PlaceCard
-                    key={place.name}
-                    name={place.name}
-                    price={place.price}
-                    type={place.type}
-                    premium={place.premium}
-                    bookmarked={place.bookmarked}
-                  />
-                ))}
-              </div>
+              <OffersList places={places} />
             </section>
             <div className="cities__right-section">
               <section className="cities__map map" />
